% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov3_fixest.R
\name{calculate_beta_jack_sparse}
\alias{calculate_beta_jack_sparse}
\title{calculate leave-one out regression coefficients}
\usage{
calculate_beta_jack_sparse(
  obj,
  cluster,
  type,
  absorb_cluster_fixef,
  return_all
)
}
\description{
calculate leave-one out regression coefficients
}
