% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeftable_summclust.R
\name{coeftable.summclust}
\alias{coeftable.summclust}
\title{Extract a coeftable for an object of type \code{summclust}}
\usage{
\method{coeftable}{summclust}(obj, ...)
}
\arguments{
\item{obj}{An object of class 'summclust'}

\item{...}{Other arguments}
}
\description{
Extract a coeftable for an object of type \code{summclust}
}
\examples{
\dontrun{
library(summclust)
library(haven)

nlswork <- read_dta("http://www.stata-press.com/data/r9/nlswork.dta")
# drop NAs at the moment
nlswork <- nlswork[, c("ln_wage", "grade", "age", "birth_yr", "union", "race", "msp", "ind_code")]
nlswork <- na.omit(nlswork)

lm_fit <- lm(
  ln_wage ~ union +  race + msp + as.factor(birth_yr) + as.factor(age) + as.factor(grade),
  data = nlswork)

res <- summclust(
   obj = lm_fit,
   cluster = ~ind_code,
   type = "CRV3"
 )

 summary(res, param = c("msp","union"))
 coeftable(res, param = c("msp","union"))
 plot(res, param = c("msp","union"))
}
}
