% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summclust_lm.R
\name{summclust.lm}
\alias{summclust.lm}
\title{Compute CR3 Jackknive variance covariance matrices of objects of type fixest}
\usage{
\method{summclust}{lm}(obj, cluster, type, ...)
}
\arguments{
\item{obj}{An object of type lm}

\item{cluster}{A clustering vector}

\item{type}{"CRV3" or "CRV3J" following MacKinnon, Nielsen & Webb}

\item{...}{other function arguments passed to 'vcov'}
}
\description{
Compute CR3 Jackknive variance covariance matrices of objects of type fixest
}
\examples{
\dontrun{
library(summclust)
library(haven)

nlswork <- read_dta("http://www.stata-press.com/data/r9/nlswork.dta")
# drop NAs at the moment
nlswork <- nlswork[, c("ln_wage", "grade", "age", "birth_yr", "union", "race", "msp", "ind_code")]
nlswork <- na.omit(nlswork)

lm_fit <- lm(
  ln_wage ~ union +  race + msp + as.factor(birth_yr) + as.factor(age) + as.factor(grade),
  data = nlswork)

res <- summclust(
   obj = lm_fit,
   cluster = ~ind_code,
   type = "CRV3"
 )

 summary(res, param = c("msp","union"))
 coeftable(res, param = c("msp","union"))
 plot(res, param = c("msp","union"))
}
}
