% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeftable_summclust.R
\name{coeftable}
\alias{coeftable}
\title{Inference based on CRV3 and CRV3J estimates}
\usage{
coeftable(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{summclust}}

\item{...}{Other arguments}
}
\description{
Inference based on CRV3 and CRV3J estimates
}
