% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeftable_summclust.R
\name{coeftable.summclust}
\alias{coeftable.summclust}
\title{Differences between lmtest::coeftest() and summclust
to compute pvalues and confidence intervals, coeftable() uses
a t(G-1) distribution
summclust uses tstat with different df, different p-val type}
\usage{
\method{coeftable}{summclust}(obj, ..., param)
}
\arguments{
\item{obj}{An object of class 'summclust'}

\item{...}{Other arguments}

\item{param}{A character scalar}
}
\description{
Differences between lmtest::coeftest() and summclust
to compute pvalues and confidence intervals, coeftable() uses
a t(G-1) distribution
summclust uses tstat with different df, different p-val type
}
