% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeftable_summclust.R
\name{coeftable.summclust}
\alias{coeftable.summclust}
\title{Differences between lmtest::coeftest() and summclust:
summclust uses tstat with different df, different p-val type}
\usage{
\method{coeftable}{summclust}(obj, param)
}
\description{
Differences between lmtest::coeftest() and summclust:
summclust uses tstat with different df, different p-val type
}
