% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summclust.R
\name{plot.summclust}
\alias{plot.summclust}
\title{plot methods for summclust objects}
\usage{
\method{plot}{summclust}(obj, coef = NULL)
}
\arguments{
\item{obj}{An object of type summclust}

\item{coef}{The coefficients for which leverages should be plotted}
}
\description{
plot methods for summclust objects
}
