% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summclust.R
\name{plot.summclust}
\alias{plot.summclust}
\title{autoplot method for summclust xs}
\usage{
\method{plot}{summclust}(x, ..., param)
}
\arguments{
\item{x}{An x of type summclust}

\item{...}{other optional function arguments}

\item{param}{Character vector. Which parameters should be plotted?}
}
\description{
autoplot method for summclust xs
}
