% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summclust.R
\name{plot.summclust}
\alias{plot.summclust}
\title{plot methods for summclust objects}
\usage{
\method{plot}{summclust}(x, ...)
}
\arguments{
\item{x}{An object of type summclust}

\item{...}{other optional function arguments}
}
\description{
plot methods for summclust objects
}
