% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_summclust.R
\name{summary.summclust}
\alias{summary.summclust}
\title{\code{summary()} method for objects of type \code{summclust}}
\usage{
\method{summary}{summclust}(object, ..., param)
}
\arguments{
\item{object}{An objet of type summclust}

\item{...}{misc arguments}

\item{param}{Character vector. The parameters to be summarized}
}
\description{
\code{summary()} method for objects of type \code{summclust}
}
