% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summclust_fixest.R
\name{summclust.fixest}
\alias{summclust.fixest}
\title{Compute CR3 Jackknive variance covariance matrices of objects
of type fixest}
\usage{
\method{summclust}{fixest}(obj, cluster, params, absorb_cluster_fixef = TRUE, type, ...)
}
\arguments{
\item{obj}{An object of type fixest}

\item{cluster}{A clustering vector}

\item{params}{A character vector of variables for which leverage statistics
should be computed. If NULL, leverage statistics will be computed for all
k model covariates}

\item{absorb_cluster_fixef}{TRUE by default. Should the cluster fixed
effects be projected out? This increases numerical stability.}

\item{type}{"CRV3" or "CRV3J" following MacKinnon, Nielsen & Webb}

\item{...}{other function arguments passed to 'vcov'}
}
\description{
Compute CR3 Jackknive variance covariance matrices of objects
of type fixest
}
\examples{
\donttest{
library(summclust)
library(haven)
library(fixest)

nlswork <- read_dta("http://www.stata-press.com/data/r9/nlswork.dta")
# drop NAs at the moment
nlswork <- nlswork[, c("ln_wage", "grade", "age", "birth_yr", "union", "race", "msp", "ind_code")]
nlswork <- na.omit(nlswork)

feols_fit <- feols(
  ln_wage ~ union +  race + msp + as.factor(birth_yr) + as.factor(age) + as.factor(grade),
  data = nlswork)

res <- summclust(
   obj = lm_fit,
   cluster = ~ind_code,
   params = c("msp", "union")
 )

 summary(res)
 coeftable(res)
 plot(res)
}
}
