% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summclust_fixest.R
\name{summclust.fixest}
\alias{summclust.fixest}
\title{Compute CR3 Jackknive variance covariance matrices of objects of type fixest}
\usage{
\method{summclust}{fixest}(obj, cluster, fe = TRUE, type, fixef.K = FALSE, ...)
}
\arguments{
\item{obj}{An object of type fixest}

\item{cluster}{A clustering vector}

\item{type}{"CRV3" or "CRV3J" following MacKinnon, Nielsen & Webb}

\item{fixef.K}{Should out-projected fixed effects be counted when computing
the number of estimated parameters? FALSE by default}

\item{...}{other function arguments passed to 'vcov'}

\item{logical}{fe TRUE by default. Should the cluster variable be projected out?
This increases numerical stability.}
}
\description{
Compute CR3 Jackknive variance covariance matrices of objects of type fixest
}
