% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summclust_fixest.R
\name{summclust.fixest}
\alias{summclust.fixest}
\title{Compute CR3 Jackknive variance covariance matrices of objects of type fixest}
\usage{
\method{summclust}{fixest}(obj, cluster, absorb_cluster_fixef = TRUE, type, ...)
}
\arguments{
\item{obj}{An object of type fixest}

\item{cluster}{A clustering vector}

\item{absorb_cluster_fixef}{TRUE by default. Should the cluster fixed
effects be projected out? This increases numerical stability.}

\item{type}{"CRV3" or "CRV3J" following MacKinnon, Nielsen & Webb}

\item{...}{other function arguments passed to 'vcov'}
}
\description{
Compute CR3 Jackknive variance covariance matrices of objects of type fixest
}
\examples{

\dontrun{
library(summclust)
library(fixest)
library(haven)

nlswork <- read_dta("http://www.stata-press.com/data/r9/nlswork.dta")
# drop NAs at the moment
nlswork <- nlswork[, c("ln_wage", "grade", "age", "birth_yr", "union", "race", "msp", "ind_code")]
nlswork <- na.omit(nlswork)

feols_fit <- feols(
  ln_wage ~ union +  race + msp | birth_yr + age + grade,
  data = nlswork,
  cluster = ~ind_code)
}
}
