% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summclust_lm.R
\name{summclust.lm}
\alias{summclust.lm}
\title{Compute CR3 Jackknive variance covariance matrices of objects of type fixest}
\usage{
\method{summclust}{lm}(obj, cluster, type, ...)
}
\arguments{
\item{obj}{An object of type lm}

\item{cluster}{A clustering vector}

\item{type}{"CRV3" or "CRV3J" following MacKinnon, Nielsen & Webb}

\item{...}{other function arguments passed to 'vcov'}
}
\description{
Compute CR3 Jackknive variance covariance matrices of objects of type fixest
}
