% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summclust.R
\name{summclust}
\alias{summclust}
\title{Compute leverage statistics for clustered inference}
\usage{
summclust(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{lm} or \code{fixest}}

\item{...}{Other arguments}
}
\description{
Compute leverage statistics for clustered inference
}
