% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov3_fixest.R
\name{vcov_CR3J.fixest}
\alias{vcov_CR3J.fixest}
\title{Compute CR3 Jackknive variance covariance matrices of objects of type fixest}
\usage{
\method{vcov_CR3J}{fixest}(
  obj,
  cluster,
  type = "CRV3",
  return_all = FALSE,
  absorb_cluster_fixef = TRUE,
  ...
)
}
\arguments{
\item{obj}{An object of type fixest}

\item{cluster}{A clustering vector}

\item{type}{"CRV3" or "CRV3J" following MacKinnon, Nielsen & Webb.
CRV3 by default}

\item{return_all}{Logical scalar, FALSE by default. Should only
the vcov be returned (FALSE) or additional results (TRUE)}

\item{absorb_cluster_fixef}{TRUE by default. Should the cluster fixed
effects be projected out? This increases numerical stability.}

\item{...}{other function arguments passed to 'vcov'}
}
\description{
Compute CR3 Jackknive variance covariance matrices of objects of type fixest
}
\examples{
\dontrun{
library(summclust)
library(haven)
library(fixest)

nlswork <- read_dta("http://www.stata-press.com/data/r9/nlswork.dta")
# drop NAs at the moment
nlswork <- nlswork[, c("ln_wage", "grade", "age", "birth_yr", "union", "race", "msp", "ind_code")]
nlswork <- na.omit(nlswork)

lm_fit <- feols(
  ln_wage ~ union +  race + msp + as.factor(birth_yr) + as.factor(age) + as.factor(grade),
  data = nlswork)

res <- summclust(
   obj = lm_fit,
   cluster = ~ind_code,
   type = "CRV3"
 )

 summary(res, param = c("msp","union"))
 coeftable(res, param = c("msp","union"))
 plot(res, param = c("msp","union"))
}
}
